+++ src/com/pablo/Tanque.java
@@ -60,7 +60,7 @@ public class Tanque extends Vehículo {
         return experiencia;
     }
 
-    public int  getHp () { return  hp;}
+    public double  getHp () { return  hp;}
 
     public int  getNivel () { return  nivel;}
 
@@ -92,11 +92,11 @@ public class Tanque extends Vehículo {
 
     public void setExperiencia(int experiencia) {this.experiencia = experiencia;}
 
-    public void setHp(int hp, int nivel) {
+    public void setHp(double hp, int nivel) {
         this.hp = hp + ((nivel-1)*50);
     }
 
-    public void setHp(int hp) {
+    public void setHp(double hp) {
         this.hp = hp;
     }
